ai_hedge_fund/requirements.txt
# Core trading
backtrader==1.9.78.123
alpaca-trade-api==3.3.2

# Data
yfinance==0.2.37
pandas==2.2.2
numpy==1.26.4
requests==2.31.0

# Options & Quant
QuantLib==1.34
scipy==1.13.1
mibian==0.1.3

# ML
scikit-learn==1.5.0
lightgbm==4.3.0
xgboost==2.0.3
ta==0.11.0
pandas-ta==0.3.14b

# Reporting
quantstats==0.0.62
matplotlib==3.9.0
jinja2==3.1.4
pyfolio-reloaded==0.9.5

# LLM
openai==1.30.0
anthropic==0.26.0

# Utilities
python-dotenv==1.0.1
loguru==0.7.2
schedule==1.2.1
pytz==2024.1
